orinium_browser::engine::html::parser

Type Alias DomTree

pub type DomTree = Tree<HtmlNodeType>;

Aliased Type

pub struct DomTree {
    pub root: Rc<RefCell<TreeNode<HtmlNodeType>>>,
}

Fields
root: Rc<RefCell<TreeNode<HtmlNodeType>>>
Implementations

impl DomTree

pub fn get_elements_by_tag_name( &self, tag_name: &str, ) -> Vec<NodeRef<HtmlNodeType>>

Returns all elements with the given tag name

pub fn get_element_by_id(&self, id: &str) -> Option<NodeRef<HtmlNodeType>>

Returns the element with the given id

pub fn get_elements_by_class_name( &self, class_name: &str, ) -> Vec<NodeRef<HtmlNodeType>>

Returns all elements that have the given class

pub fn inner_text(node: &NodeRef<HtmlNodeType>) -> String

Returns the concatenated text content of this node (including children)

pub fn set_text_content(node: &NodeRef<HtmlNodeType>, new_text: &str)

Replace all text content of this node with the given string

pub fn collect_text_by_tag(&self, tag_name: &str) -> Vec<String>

指定したタグ名の要素のテキストノードをすべて集める